move shannonCombo classes in a separate file
