+++ src/main/java/it/smartphonecombo/uecapabilityparser/model/shannon/ShannonNrUECap.kt
@@ -3,9 +3,7 @@
 package it.smartphonecombo.uecapabilityparser.model.shannon
 
 import it.smartphonecombo.uecapabilityparser.extension.Band
-import it.smartphonecombo.uecapabilityparser.model.BCS
 import it.smartphonecombo.uecapabilityparser.model.BwClass
-import it.smartphonecombo.uecapabilityparser.model.PowerClass
 import it.smartphonecombo.uecapabilityparser.model.modulation.ModulationOrder
 import kotlin.math.max
 import kotlinx.serialization.ExperimentalSerializationApi
@@ -37,86 +35,6 @@ data class ShannonNrUECap(
 )
 
 @Serializable
-data class ComboGroup(
-    /** Some features that applies to the whole combo group. */
-    @ProtoNumber(1) val comboFeatures: ComboFeatures,
-    /** List of combos that share the same [ComboFeatures]. */
-    @ProtoNumber(2) val combos: List<ShannonCombo> = emptyList(),
-)
-
-@Serializable
-data class ComboFeatures(
-    /**
-     * The supportedBandwidthCombinationSet that applies to the Nr Components.
-     *
-     * It's stored as a 32bit unsigned int, each of its bits has the same value of the corresponding
-     * bit in the BitString.
-     */
-    @ProtoNumber(1) private val rawBcsNr: UInt = 0u,
-
-    /**
-     * The supportedBandwidthCombinationSet that applies to the IntraEnDc Components
-     * (supportedBandwidthCombinationSetIntraENDC).
-     *
-     * It's stored as a 32bit unsigned int, each of its bits has the same value of the corresponding
-     * bit in the BitString.
-     */
-    @ProtoNumber(2) private val rawBcsIntraEndc: UInt = 0u,
-
-    /**
-     * The supported Bandwidth Combination Set that applies to the Eutra Components
-     * (supportedBandwidthCombinationSetEUTRA-v1530).
-     *
-     * It's stored as a 32bit unsigned int, each of its bits has the same value of the corresponding
-     * bit in the BitString.
-     */
-    @ProtoNumber(3) private val rawBcsEutra: UInt = 0u,
-
-    /**
-     * Power Class of the whole combination, it's stored as an enum.
-     *
-     * Note that this doesn't override the powerclass of the uplink bands.
-     *
-     * For FR1 0 -> Default, 1 -> PC2, 2 -> PC1.5
-     *
-     * For FR2 0 -> Default
-     */
-    @ProtoNumber(4) private val rawPowerClass: Int = 0,
-
-    /**
-     * intraBandENDC-Support is stored as an enum.
-     *
-     * 0 -> contiguous, 1 -> non-contiguous, 2 -> both.
-     */
-    @ProtoNumber(5) private val rawIntraBandEnDcSupport: Int = 0
-) {
-    val bcsNr
-        get() = BCS.fromBinaryString(rawBcsNr.toString(2))
-
-    val bcsIntraEndc
-        get() = BCS.fromBinaryString(rawBcsIntraEndc.toString(2))
-
-    val bcsEutra
-        get() = BCS.fromBinaryString(rawBcsIntraEndc.toString(2))
-
-    val powerClass
-        get() =
-            when (rawPowerClass) {
-                1 -> PowerClass.PC2
-                2 -> PowerClass.PC1dot5
-                else -> PowerClass.NONE
-            }
-}
-
-@Serializable
-data class ShannonCombo(
-    /** List of Components. */
-    @ProtoNumber(1) val components: List<ShannonComponent> = emptyList(),
-    /** A bit mask stored as unsigned int that enables or disables some features. */
-    @ProtoNumber(2) val bitMask: UInt
-)
-
-@Serializable
 data class ShannonComponent(
     /**
      * LTE Bands are stored as they are.
@@ -133,8 +51,8 @@ data class ShannonComponent(
 
     /**
      * For LTE this is FeatureSetEUTRA-DownlinkId, the corresponding FeatureSetDL-r15 seems to be
-     * hardcoded elsewhere (see [ShannonFeatureSetEutra]). Note that the index starts from
-     * 1 as per 3GPP spec, 0 means DL not supported.
+     * hardcoded elsewhere (see [ShannonFeatureSetEutra]). Note that the index starts from 1 as per
+     * 3GPP spec, 0 means DL not supported.
      *
      * For NR this sets some features that applies to the whole component (not PerCC). Empirically 1
      * -> FR1, 2 -> FR2.
@@ -143,8 +61,8 @@ data class ShannonComponent(
 
     /**
      * For LTE this is FeatureSetEUTRA-UplinkId, the corresponding FeatureSetUL-r15 seems to be
-     * hardcoded elsewhere (see [ShannonFeatureSetEutra]). Note that the index starts from
-     * 1 as per 3GPP spec, 0 means UL not supported.
+     * hardcoded elsewhere (see [ShannonFeatureSetEutra]). Note that the index starts from 1 as per
+     * 3GPP spec, 0 means UL not supported.
      *
      * For NR this sets some features that applies to the whole component (not PerCC). Empirically 1
      * -> SRS ports per resource 1, 2 -> SRS ports per resource 2.
